+++ types/elem_fixed_size_series.go
@@ -9,18 +9,18 @@ import (
 	"unsafe"
 )
 
-func EncodeFixedSeries(encFn EncoderFn, length uint32, elemMemSize uintptr, eb *EncodingBuffer, p unsafe.Pointer) {
+func EncodeFixedSeries(encFn EncoderFn, length uint64, elemMemSize uintptr, eb *EncodingBuffer, p unsafe.Pointer) {
 	memOffset := uintptr(0)
-	for i := uint32(0); i < length; i++ {
+	for i := uint64(0); i < length; i++ {
 		elemPtr := unsafe.Pointer(uintptr(p) + memOffset)
 		memOffset += elemMemSize
 		encFn(eb, elemPtr)
 	}
 }
 
-func DecodeFixedSeries(decFn DecoderFn, length uint32, elemMemSize uintptr, dr *DecodingReader, p unsafe.Pointer) error {
+func DecodeFixedSeries(decFn DecoderFn, length uint64, elemMemSize uintptr, dr *DecodingReader, p unsafe.Pointer) error {
 	memOffset := uintptr(0)
-	for i := uint32(0); i < length; i++ {
+	for i := uint64(0); i < length; i++ {
 		elemPtr := unsafe.Pointer(uintptr(p) + memOffset)
 		memOffset += elemMemSize
 		if err := decFn(dr, elemPtr); err != nil {
@@ -30,7 +30,7 @@ func DecodeFixedSeries(decFn DecoderFn, length uint32, elemMemSize uintptr, dr *
 	return nil
 }
 
-func DecodeFixedSlice(decFn DecoderFn, elemLen uint32, bytesLen uint32, limit uint32, alloc ptrutil.SliceAllocationFn, elemMemSize uintptr, dr *DecodingReader, p unsafe.Pointer) error {
+func DecodeFixedSlice(decFn DecoderFn, elemLen uint64, bytesLen uint64, limit uint64, alloc ptrutil.SliceAllocationFn, elemMemSize uintptr, dr *DecodingReader, p unsafe.Pointer) error {
 	if elemLen == 0 {
 		return errors.New("cannot read a dynamic-length series of 0-length elements")
 	}
