+++ types/ssz_bytes_n.go
@@ -12,7 +12,7 @@ import (
 )
 
 type SSZBytesN struct {
-	length uint32
+	length uint64
 }
 
 func NewSSZBytesN(typ reflect.Type) (*SSZBytesN, error) {
@@ -23,20 +23,20 @@ func NewSSZBytesN(typ reflect.Type) (*SSZBytesN, error) {
 		return nil, fmt.Errorf("typ is not a bytes array")
 	}
 	length := typ.Len()
-	res := &SSZBytesN{length: uint32(length)}
+	res := &SSZBytesN{length: uint64(length)}
 	return res, nil
 }
 
-func (v *SSZBytesN) FuzzReqLen() uint32 {
+func (v *SSZBytesN) FuzzReqLen() uint64 {
 	return v.length
 }
 
-func (v *SSZBytesN) FixedLen() uint32 {
+func (v *SSZBytesN) FixedLen() uint64 {
 	// 1 byte per element, just the same as the length
 	return v.length
 }
 
-func (v *SSZBytesN) MinLen() uint32 {
+func (v *SSZBytesN) MinLen() uint64 {
 	// 1 byte per element, just the same as the length
 	return v.length
 }
@@ -62,7 +62,7 @@ func (v *SSZBytesN) HashTreeRoot(h HashFn, p unsafe.Pointer) [32]byte {
 	sh := ptrutil.GetSliceHeader(p, v.length)
 	data := *(*[]byte)(unsafe.Pointer(sh))
 	leafCount := (v.length + 31) >> 5
-	leaf := func(i uint32) []byte {
+	leaf := func(i uint64) []byte {
 		s := i << 5
 		e := (i + 1) << 5
 		// pad the data
